kola/tests/iso: fold testiso iso-fips.uefi test
Migrate the last remaining test from the legacy 'kola testiso'
command into the standard 'kola run' This completes the migration
of all ISO tests and removes the now-obsolete testiso.go file.
